Move pose_graph_stub/trajectory_builder_stub to internal.
